internal/calculator/calculator_test.go
package calculator

import (
	"github.com/sv345922/arithmometer_v2/internal/entities"
	"testing"
)

func TestDo(t *testing.T) {
	type args struct {
		c *entities.MessageTask
	}
	tests := []struct {
		name    string
		args    args
		want    float64
		wantErr bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			got, err := Do(tt.args.c)
			if (err != nil) != tt.wantErr {
				t.Errorf("Parse() error = %v, wantErr %v", err, tt.wantErr)
				return
			}
			if got != tt.want {
				t.Errorf("Parse() got = %v, want %v", got, tt.want)
			}
		})
	}
}
